[state-refactoring] Prop Cleanup
